splitted profiles keeper
